Exchanged BigInt with java.math.BigInteger

+++ src/de.hpi.swa.graal.squeak/src/de/hpi/swa/graal/squeak/image/SqueakImageContext.java
@@ -2,6 +2,7 @@ package de.hpi.swa.graal.squeak.image;
 
 import java.io.File;
 import java.io.PrintWriter;
+import java.math.BigInteger;
 import java.nio.channels.SeekableByteChannel;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -53,7 +54,6 @@ import de.hpi.swa.graal.squeak.nodes.primitives.PrimitiveNodeFactory;
 import de.hpi.swa.graal.squeak.shared.SqueakLanguageConfig;
 import de.hpi.swa.graal.squeak.util.ArrayConversionUtils;
 import de.hpi.swa.graal.squeak.util.ArrayUtils;
-import de.hpi.swa.graal.squeak.util.BigInt;
 import de.hpi.swa.graal.squeak.util.FrameAccess;
 import de.hpi.swa.graal.squeak.util.InterruptHandlerState;
 import de.hpi.swa.graal.squeak.util.MiscUtils;
@@ -371,7 +371,7 @@ public final class SqueakImageContext {
         return l;
     }
 
-    public AbstractSqueakObject wrap(final BigInt i) {
+    public AbstractSqueakObject wrap(final BigInteger i) {
         return new LargeIntegerObject(this, i);
     }
